+++ src/config/DefaultConfigGetter.ts
@@ -1,0 +1,7 @@
+export interface DefaultConfigGetter<C> {
+
+    /**
+     * get the default config block entry for the target
+     */
+    getDefaultConfig(): Partial<C> & { [index: string]: any };
+}