+++ src/target/lua/LuaDefaultConfig.ts
@@ -1,0 +1,11 @@
+import { DefaultConfigGetter } from "../../config/DefaultConfigGetter";
+import { LuaConfig } from "./LuaConfig";
+
+export class LuaDefaultConfig implements DefaultConfigGetter<LuaConfig> {
+
+    public getDefaultConfig(): LuaConfig {
+
+        // currently no default config for lua
+        return {};
+    }
+}