+++ src/target/wow/special/WowSpecial.ts
@@ -1,4 +1,5 @@
 import { WowPathBuilder } from "./WowPathBuilder";
 import { WowPostTranspile } from "./WowPostTranspile";
+import { WowGlobalLibrary } from "./WowGlobalLibrary";
 
-export interface WowSpecial extends WowPathBuilder, WowPostTranspile { }
+export interface WowSpecial extends WowPathBuilder, WowPostTranspile, WowGlobalLibrary { }