+++ src/target/wow/special/index.ts
@@ -1,5 +1,6 @@
 export * from "./WowPathBuilder";
 export * from "./WowPostTranspile";
+export * from "./WowGlobalLibrary";
 
 // bundle interface
 export * from "./WowSpecial";