+++ test/target/wow/WowImportDeclaration.spec.ts
@@ -17,9 +17,11 @@ import { WowConfig } from "../../../src/target/wow/WowConfig";
         this.runCodeAndExpectResult("wow", [{
             code: `import * as a from "./b"`,
             expected: [
-                `local a = ${WowKeywords.IMPORT_LIB_NAME}.get("${this.lastProject.name}@b")`
+                `local a = ${WowKeywords.IMPORT_LIB_NAME}.get("b")`
             ]
-        }])
+        }], {
+            personalizedLibrary: false
+        })
     }
 
     @test "Names imports"() {
@@ -28,17 +30,19 @@ import { WowConfig } from "../../../src/target/wow/WowConfig";
             {
                 code: `import {A} from "./a"`,
                 expected: [
-                    `local A = ${WowKeywords.IMPORT_LIB_NAME}.get("${this.lastProject.name}@a").A`
+                    `local A = ${WowKeywords.IMPORT_LIB_NAME}.get("a").A`
                 ]
 
             }, {
                 code: `import { A as B } from "./a";`,
                 expected: [
-                    `local B = ${WowKeywords.IMPORT_LIB_NAME}.get("${this.lastProject.name}@a").A`
+                    `local B = ${WowKeywords.IMPORT_LIB_NAME}.get("a").A`
                 ]
 
             }
-        ]);
+        ], {
+            personalizedLibrary: false
+        });
 
     }
 
@@ -48,13 +52,15 @@ import { WowConfig } from "../../../src/target/wow/WowConfig";
             {
                 code: `import {A, B, C as D} from "./a"`,
                 expected: [
-                    `local A = ${WowKeywords.IMPORT_LIB_NAME}.get("${this.lastProject.name}@a").A`,
-                    `local B = ${WowKeywords.IMPORT_LIB_NAME}.get("${this.lastProject.name}@a").B`,
-                    `local D = ${WowKeywords.IMPORT_LIB_NAME}.get("${this.lastProject.name}@a").C`
+                    `local A = ${WowKeywords.IMPORT_LIB_NAME}.get("a").A`,
+                    `local B = ${WowKeywords.IMPORT_LIB_NAME}.get("a").B`,
+                    `local D = ${WowKeywords.IMPORT_LIB_NAME}.get("a").C`
                 ]
 
             }
-        ]);
+        ], {
+            personalizedLibrary: false
+        });
 
     }
 
@@ -63,9 +69,11 @@ import { WowConfig } from "../../../src/target/wow/WowConfig";
         this.runCodeAndExpectResult("wow",[{
             code: `import {A} from "./a/b/c/d";`,
             expected: [
-                `local A = ${WowKeywords.IMPORT_LIB_NAME}.get("${this.lastProject.name}@a/b/c/d").A`
+                `local A = ${WowKeywords.IMPORT_LIB_NAME}.get("a/b/c/d").A`
             ]
-        }])
+        }], {
+            personalizedLibrary: false
+        })
     }
 
     @test "File imports should fail"() {